Move CodeReader.navigateToToken into PanelEditor
